requirements.txt
Flask==2.3.3
boto3==1.28.44
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.45
pymysql==1.1.0
python-dotenv==1.1.1